Add files via upload
I am delighted to share this Letter of Recommendation from my recent internship as a Blockchain, Crypto, and Smart Contract Intern at Future Interns.

It was a privilege to contribute to challenging projects, and I'm grateful for the team's positive feedback on my "exceptional performance," "strong work ethic," and "ability to excel". My sincere thanks to the Future Interns team for this endorsement and for an unforgettable learning experience.

Looking forward to applying what I've learned in future endeavors.